+++ src/home/index.js
@@ -26,15 +26,16 @@ import {
   fetchData, isMobile, getOSMTiles,
   isURL, theme, updateHistory,
 } from '../utils/utils';
-import { ICONLIMIT } from '../Constants';
+import { ICONLIMIT, NO_LEGEND_LAYERS_REGEX } from '../Constants';
 import DeckSidebarContainer from
   '../components/decksidebar/DeckSidebarContainer';
 
 import '../App.css';
 import Tooltip from '../components/tooltip';
-import { isObject } from '../utils/JSUtils';
+import { isArray, isObject } from '../utils/JSUtils';
 import { generateLayer, initViewState,
   getViewPort } from './util';
+import { jsonFromKeySetObject } from '../utils/api';
 
 // Set your mapbox access token here
 const MAPBOX_ACCESS_TOKEN = process.env.REACT_APP_MAPBOX_ACCESS_TOKEN;
@@ -81,9 +82,9 @@ export default class Home extends React.Component {
     this._initDataState = this._initDataState.bind(this);
     this._updateStateAndLayers = this._updateStateAndLayers.bind(this);
     this._resize = this._resize.bind(this);
-    this._updateURL = this._updateURL.bind(this);
-    // TODO: can let user change the 300
-    this._throttleUR = throttle((v) => this._updateURL(v), 300);
+    this._updateHistory = this._updateHistory.bind(this);
+    // TODO: expose interval
+    this._throttleUR = throttle((v) => this._updateHistory(v), 300);
     this._urlCallback = this._urlCallback.bind(this);
   }
 
@@ -91,14 +92,23 @@ export default class Home extends React.Component {
     // props change
     const { data, defaultURL, geographyURL,
       geographyColumn } = nextProps;
-    if (JSON.stringify(data) !== JSON.stringify(this.props.data) ||
-      defaultURL !== this.props.defaultURL ||
-      geographyURL !== this.props.geographyURL ||
-      geographyColumn !== this.props.geographyColumn) {
+    if(!isObject(data)) return
+    const r = isArray(data.features)
+      && Math.floor(Math.random() * data.features.length)
+    /**
+     * If user has added data using Add data button
+     * ignore props updates
+     */
+    if (!this.state.reset && this.props.data &&
+      ((isArray(this.props.data.features)
+        && JSON.stringify(data.features[r]) !==
+        JSON.stringify(this.props.data.features[r])) ||
+        defaultURL !== this.props.defaultURL ||
+        geographyURL !== this.props.geographyURL ||
+        geographyColumn !== this.props.geographyColumn)) {
+      this.map && this.map.stop()
       this._initDataState()
-      return true
     }
-    //TODO: return false?
   }
 
   componentDidMount() {
@@ -239,10 +249,17 @@ export default class Home extends React.Component {
 
   _callGenerateLayer(values = {}) {
     const updateState = generateLayer(
-      values, this.props, this.state, this._renderTooltip,
+      values, this.state, this._renderTooltip,
       this._callGenerateLayer
     )
-    updateState && this.setState({ ...updateState })
+    if(isObject(updateState)) {
+      this.setState({ ...updateState })
+      // TODO: send this to a factory
+      const { multiVarSelect } = updateState;
+      const { onStateChange } = this.props;
+      typeof onStateChange === 'function'
+        && onStateChange({select: jsonFromKeySetObject(multiVarSelect)})
+    }
   }
 
   _fitViewport(newData, bboxLonLat) {
@@ -286,8 +303,9 @@ export default class Home extends React.Component {
     }
   }
 
-  _updateURL(viewport) {
+  _updateHistory(viewport) {
     const { subsetBoundsChange, lastViewPortChange } = this.state;
+    const { onViewStateChange } = this.props;
 
     //if we do history.replace/push 100 times in less than 30 secs
     // browser will crash
@@ -301,6 +319,9 @@ export default class Home extends React.Component {
         }
       });
       this.setState({ lastViewPortChange: new Date() })
+      // TODO: in future send this to factory of callbacks
+      typeof onViewStateChange === 'function'
+        && onViewStateChange({viewState: viewport})
     }
 
     if (subsetBoundsChange) {
@@ -320,8 +341,7 @@ export default class Home extends React.Component {
       loading, mapStyle, alert, data, filtered, bottomPanel,
       layerName, legend, coords } = this.state;
     const showLegend = legend
-      && !new RegExp("grid|sgrid|text|heatmap|icon", "i").test(layerName)
-
+      && !new RegExp(NO_LEGEND_LAYERS_REGEX, "i").test(layerName)
     return (
       <div>
         <div className="loader" style={{
@@ -379,17 +399,32 @@ export default class Home extends React.Component {
           </DeckGL>
         </MapGL>
         {!hideSidebar && <DeckSidebarContainer
+          key="decksidebar"
+          /** data */
+          unfilteredData={data && data.features}
+          data={filtered}
+          /** api */
           hideCharts={hideCharts}
-          map={this.map} deck={this.deck}
           hideChartGenerator={hideChartGenerator}
           leftSidebarContent={leftSidebarContent}
           dark={dark}
           layerName={layerName}
+          layerOptions={this.state.layerOptions}
+          /** UI */
+          subsetBoundsChange={this.state.subsetBoundsChange}
+          // TODO: generalise datasetName
+          datasetName={defaultURL}
+          bottomPanel={bottomPanel}
+          map={this.map} deck={this.deck}
           isMobile={isMobile()}
-          key="decksidebar"
           alert={alert}
-          unfilteredData={data && data.features}
-          data={filtered}
+          // only during first load
+          // DeckSidebar ignores this prop later
+          multiVarSelect={this.state.multiVarSelect}
+          /** callbacks */
+          onlocationChange={(bboxLonLat) => {
+            this._fitViewport(undefined, bboxLonLat)
+          }}
           colourCallback={(colorName) => {
             this._callGenerateLayer({ cn: colorName })
           }}
@@ -405,22 +440,14 @@ export default class Home extends React.Component {
             })
             this._callGenerateLayer()
           }}
-          subsetBoundsChange={this.state.subsetBoundsChange}
-          onlocationChange={(bboxLonLat) => {
-            this._fitViewport(undefined, bboxLonLat)
-          }}
-          // TODO: generalise datasetName
-          datasetName={defaultURL}
-          bottomPanel={bottomPanel}
-          // only during first load
-          // DeckSidebar ignores this prop later
-          multiVarSelect={this.state.multiVarSelect}
         />}
         {
           showLegend &&
           <div
             id="tgve-legend"
             style={{
+              overflowY: 'scroll',
+              maxHeight: '70%',
               ...theme(this.props.dark)
             }}
             className="right-side-panel mapbox-legend">
@@ -446,8 +473,9 @@ export default class Home extends React.Component {
    * @param {String} geoColumn the matching column name
    * for `geojson_returned` and `geography_returned` params
    */
-  _urlCallback(url_returned, geojson_returned,
-    geography_returned, geoColumn) {
+  _urlCallback(params = {}) {
+    const {geojson_returned,
+      geography_returned, geoColumn, reset} = params;
     this.setState({
       /**
        * This set state can take care of all
@@ -457,12 +485,10 @@ export default class Home extends React.Component {
        * `this._fitViewport(geojson_returned)` &&
        * `generateLayer` take care of it.
        *
-       * 2. if a URL has been returned,
-       *
-       * 3. if we are resetting, that means start from
+       * 2. if we are resetting, that means start from
        * fresh: this._initDataState
        *
-       * 4. if (1) is the case but geojson
+       * 3. if (1) is the case but geojson
        * is invalid or corrupt, then do not
        * update state's `data` or `geography`
        * and fail on the try below.
@@ -471,48 +497,40 @@ export default class Home extends React.Component {
       tooltip: "",
       loading: true,
       coords: null,
-    })
-    if (geojson_returned) {
-      // confirm valid geojson
-      try {
-        // do not move this setState up
-        // as data returned could be
-        // corrupt
-        this.setState({
-          data: geojson_returned,
-          geography: geography_returned || null,
-          geographyColumn: geoColumn,
-        }, () => {
-          // go with geography first fallback onto data source.
-          this._fitViewport(geography_returned || geojson_returned);
-          this._callGenerateLayer()
-        })
-      } catch (error) {
-        // load up default
-        this._fetchAndUpdateState(undefined,
-          { content: error.message });
-      }
-    } else {
-      if (isURL(url_returned)) {
-        fetchData(url_returned, (data, error) => {
-          if (!error) {
-            this._updateStateAndLayers(
-              // geoErr, geojson, data, customError, geographyURL
-              false, null, data
-            )
-          } else {
-            this.setState({
-              loading: false,
-              alert: { content: 'Could not reach: ' + url_returned }
-            });
-          }
-        })
+      multiVarSelect: {},
+      geography: null,
+      data: reset ? [] : this.state.data,
+      filtered: reset ? [] : this.state.filtered,
+      layers: reset ? [] : this.state.layers,
+      legend: reset ? null : this.state.legend,
+      reset
+    }, () => {
+      if (geojson_returned) {
+        // confirm valid geojson
+        try {
+          // do not move this setState up
+          // as data returned could be
+          // corrupt
+          this.setState({
+            data: geojson_returned,
+            geography: geography_returned,
+            geographyColumn: geoColumn,
+          }, () => {
+            // go with geography first fallback onto data source.
+            this._fitViewport(geography_returned || geojson_returned);
+            this._callGenerateLayer()
+          })
+        } catch (error) {
+          // load up default
+          this.setState({alert: { content: error.message }})
+          this._initDataState();
+        }
       } else {
         // empty, so might be resetting
         // current geography and defaulturl
         this._initDataState()
       }
-    }
+    })
   }
 
   _resize() {